Add support for X11 backend on MacOS (via XQuartz)
GLFW assumes that the only backend available in MacOS is Cocoa. However, by installing XQuartz, you can also build X11 apps. This commit makes the changes needed for GLFW to be able to build for the X11 backend on MacOS. Note that no changes were made to the build system, as I build GLFW with a custom Makefile rather than with CMake. Also note that in order for the app to find the X11 libraries, you need to run the app executable prepending DYLD_FALLBACK_LIBRARY_PATH="/opt/X11/lib" to it